Doors Paintable
Doors are now Paintable.
Someone requested it on the workshop page, and I can't really think of any reason not to have them paintable. Also, it was added into Core in 1.4 something, so I imagine it just wasn't considered when the mod was first made. I did not make the remote and lever paintable however, as it did not really look all that great. Everything else seems to look fine.